Add a custom implementation of binary search and exponential search to improve KC algorithm performance
The exponential search algorithm is used because we generally expect to find matches near the beginning of the search range and a custom implementation of binary search that makes it easier to avoid some extra instructions and bounds checks since we don't care about exact matches.

Based on our benchmarks this offers a 10-30% speedup.