Accellerate the KC algorithm using a cached list of offsets in thresh
This more or less addresses the `R` term in the complexity since multiple visits to the same match list entry require less work on each iteration.

Add more discussion about the tradeoffs here.  And fix a few small issues in the benchmarks which now reveal the insights of meyers more directly.