Add an implementation of the O(NP) Wu-Manber-Meyers LCS algorithm
This algorithm is an evolution of the Meyers O(ND) algorithm that restricts the search to roughly half as many points and uses half as much memory, in general this
leads to it executing twice as fast.

For now this is only the `_length_` computation, future work will add a full LCS computation.